extract syntax highlighting module
