+++ app/src/ui/diff/index.tsx
@@ -1,7 +1,6 @@
 import { clipboard } from 'electron'
 import * as React from 'react'
 import * as ReactDOM from 'react-dom'
-import * as Path from 'path'
 import { Disposable } from 'event-kit'
 
 import { assertNever } from '../../lib/fatal-error'
@@ -29,7 +28,6 @@ import {
   IDiff,
   IImageDiff,
   ITextDiff,
-  DiffLine,
   DiffLineType,
   ILargeTextDiff,
 } from '../../models/diff'
@@ -52,19 +50,19 @@ import { fatalError } from '../../lib/fatal-error'
 
 import { RangeSelectionSizePixels } from './edge-detection'
 import { relativeChanges } from './changed-range'
-import { getPartialBlobContents } from '../../lib/git/show'
-import { readPartialFile } from '../../lib/file-system'
 
 import { DiffSyntaxMode, IDiffSyntaxModeSpec } from './diff-syntax-mode'
-import { highlight } from '../../lib/highlighter/worker'
-import { ITokens } from '../../lib/highlighter/types'
+
+import {
+  getLineFilters,
+  getFileContents,
+  highlightContents,
+} from './syntax-highlighting'
+
 import { Button } from '../lib/button'
 
 /** The longest line for which we'd try to calculate a line diff. */
 const MaxIntraLineDiffStringLength = 4096
-
-/** The maximum number of bytes we'll process for highlighting. */
-const MaxHighlightContentLength = 256 * 1024
 
 // This is a custom version of the no-newline octicon that's exactly as
 // tall as it needs to be (8px) which helps with aligning it on the line.
@@ -78,186 +76,6 @@ const narrowNoNewlineSymbol = new OcticonSymbol(
 const NoDiffImage = encodePathAsUrl(__dirname, 'static/ufo-alert.svg')
 
 type ChangedFile = WorkingDirectoryFileChange | CommittedFileChange
-
-interface ILineFilters {
-  readonly oldLineFilter: Array<number>
-  readonly newLineFilter: Array<number>
-}
-
-interface IFileContents {
-  readonly file: ChangedFile
-  readonly oldContents: Buffer
-  readonly newContents: Buffer
-}
-
-interface IFileTokens {
-  readonly oldTokens: ITokens
-  readonly newTokens: ITokens
-}
-
-async function getOldFileContent(
-  repository: Repository,
-  file: ChangedFile
-): Promise<Buffer> {
-  if (file.status === AppFileStatus.New) {
-    return new Buffer(0)
-  }
-
-  let commitish
-
-  if (file instanceof WorkingDirectoryFileChange) {
-    // If we pass an empty string here we get the contents
-    // that are in the index. But since we call diff with
-    // --no-index (see diff.ts) we need to look at what's
-    // actually committed to get the appropriate content.
-    commitish = 'HEAD'
-  } else if (file instanceof CommittedFileChange) {
-    commitish = `${file.commitish}^`
-  } else {
-    return assertNever(file, 'Unknown file change type')
-  }
-
-  return getPartialBlobContents(
-    repository,
-    commitish,
-    file.oldPath || file.path,
-    MaxHighlightContentLength
-  )
-}
-
-async function getNewFileContent(
-  repository: Repository,
-  file: ChangedFile
-): Promise<Buffer> {
-  if (file.status === AppFileStatus.Deleted) {
-    return new Buffer(0)
-  }
-
-  if (file instanceof WorkingDirectoryFileChange) {
-    return readPartialFile(
-      Path.join(repository.path, file.path),
-      0,
-      MaxHighlightContentLength - 1
-    )
-  } else if (file instanceof CommittedFileChange) {
-    return getPartialBlobContents(
-      repository,
-      file.commitish,
-      file.path,
-      MaxHighlightContentLength
-    )
-  }
-
-  return assertNever(file, 'Unknown file change type')
-}
-
-async function getFileContents(
-  repo: Repository,
-  file: ChangedFile,
-  lineFilters: ILineFilters
-): Promise<IFileContents> {
-  const oldContentsPromise = lineFilters.oldLineFilter.length
-    ? getOldFileContent(repo, file)
-    : Promise.resolve(new Buffer(0))
-
-  const newContentsPromise = lineFilters.newLineFilter.length
-    ? getNewFileContent(repo, file)
-    : Promise.resolve(new Buffer(0))
-
-  const [oldContents, newContents] = await Promise.all([
-    oldContentsPromise.catch(e => {
-      log.error('Could not load old contents for syntax highlighting', e)
-      return new Buffer(0)
-    }),
-    newContentsPromise.catch(e => {
-      log.error('Could not load new contents for syntax highlighting', e)
-      return new Buffer(0)
-    }),
-  ])
-
-  return { file, oldContents, newContents }
-}
-
-/**
- * Figure out which lines we need to have tokenized in
- * both the old and new version of the file.
- */
-function getLineFilters(diff: ITextDiff): ILineFilters {
-  const oldLineFilter = new Array<number>()
-  const newLineFilter = new Array<number>()
-
-  const diffLines = new Array<DiffLine>()
-
-  let anyAdded = false
-  let anyDeleted = false
-
-  for (const hunk of diff.hunks) {
-    for (const line of hunk.lines) {
-      anyAdded = anyAdded || line.type === DiffLineType.Add
-      anyDeleted = anyDeleted || line.type === DiffLineType.Delete
-      diffLines.push(line)
-    }
-  }
-
-  for (const line of diffLines) {
-    // So this might need a little explaining. What we're trying
-    // to achieve here is if the diff contains only additions or
-    // only deletions we'll source all the highlighted lines from
-    // either the before or after file. That way we can completely
-    // disregard loading, and highlighting, the other version.
-    if (line.oldLineNumber !== null && line.newLineNumber !== null) {
-      if (anyAdded && !anyDeleted) {
-        newLineFilter.push(line.newLineNumber - 1)
-      } else {
-        oldLineFilter.push(line.oldLineNumber - 1)
-      }
-    } else {
-      // If there's a mix (meaning we'll have to read from both
-      // anyway) we'll prioritize the old version since
-      // that's immutable and less likely to be the subject of a
-      // race condition when someone rapidly modifies the file on
-      // disk.
-      if (line.oldLineNumber !== null) {
-        oldLineFilter.push(line.oldLineNumber - 1)
-      } else if (line.newLineNumber !== null) {
-        newLineFilter.push(line.newLineNumber - 1)
-      }
-    }
-  }
-
-  return { oldLineFilter, newLineFilter }
-}
-
-async function highlightContents(
-  contents: IFileContents,
-  tabSize: number,
-  lineFilters: ILineFilters
-): Promise<IFileTokens> {
-  const { file, oldContents, newContents } = contents
-
-  const [oldTokens, newTokens] = await Promise.all([
-    highlight(
-      oldContents.toString('utf8'),
-      Path.extname(file.oldPath || file.path),
-      tabSize,
-      lineFilters.oldLineFilter
-    ).catch(e => {
-      log.error('Highlighter worked failed for old contents', e)
-      return {}
-    }),
-    highlight(
-      newContents.toString('utf8'),
-      Path.extname(file.path),
-      tabSize,
-      lineFilters.newLineFilter
-    ).catch(e => {
-      log.error('Highlighter worked failed for new contents', e)
-      return {}
-    }),
-  ])
-
-  return { oldTokens, newTokens }
-}
 
 /**
  * Checks to see if any key parameters in the props object that are used
